+++ src/DummyLib/States/GSUITest3.h
@@ -30,7 +30,7 @@ class GSUITest3 : public GSBaseState {
     uint64_t last_frame_time_ = 0;
     double cur_fps_ = 0.0;
 
-    uint32_t click_time_ = 0;
+    uint64_t click_time_ = 0;
 
     std::shared_ptr<Gui::BitmapFont> book_main_font_, book_emph_font_, book_caption_font_;
     float test_time_counter_s = 0.0f;
