+++ src/DummyLib/States/GSUITest4.h
@@ -31,7 +31,7 @@ class GSUITest4 final : public GSBaseState {
     uint64_t last_frame_time_ = 0;
     double cur_fps_ = 0.0;
 
-    uint32_t click_time_ = 0;
+    uint64_t click_time_ = 0;
 
     std::shared_ptr<Gui::BitmapFont> dialog_font_;
     float test_time_counter_s = 0.0f;
